Type TMission::value0c as float; remove reinterpret_cast punning at 8 call sites
value0c is written as a float score at every meaningful call site across
TDefendProvinceMission, TBlockadePortMission, TScatteredShipsMission,
TAttackProvinceMission, TEscortMission, and TControlSeaZoneMission (only
ctor-init and raw byte serialization ever touched it as int/bytes). Typing
the field directly as float removes the *reinterpret_cast<float*>(&value0c)
punning at every write site and the read-site pun in
CompareMissionOrderEntriesByPriorityScore, matching the Type-modeling
guardrail (a reinterpret_cast to reach a value is a mismodelled-type smell).

Also replaced TMission::WriteTo's raw reinterpret_cast<char*>(this) + offset
byte walk with the already-named typed fields (nationId04/state08/value0c/
flag10/pathMarker06/marker11) it was reaching by hand.

Verified via `just stats`: zero score deltas on every touched function
(bit-identical codegen), all gates pass.

+++ include/game/TMission.h
@@ -42,7 +42,7 @@ public:
   short pathMarker06;    // 0x06 path/dispatch marker (set 0xffff)
   unsigned char state08; // 0x08 lifecycle state byte (ctor = 2)
   unsigned char padding09[3];
-  int value0c;            // 0x0c cached score/value (ctor = 0)
+  float value0c;          // 0x0c cached score/value (ctor = 0.0f)
   unsigned char flag10;   // 0x10 dispatch flag (SetMissionField10FromArgSlot94)
   unsigned char marker11; // 0x11 status byte (ctor = 0xff)
   unsigned char padding12[2];
